Move subscription_ack to bottom of switch cases; revised output
The `subscription_ack` message only happens once when first connecting to
the websocket server. By having it be the first case block in the switch
statement, we were checking the case unecessarily frequently given it is
basically the least common of all scenarios.

Moving it to the end of the cases list, we can reduce the amount we're
checking this case during regular execution.